refactor: 콜백 구성 재배치
